fix #369 remove condition delayMs
As I can see delayMs is the delay between the handling of the first byte
of a GPS-Message block and the last Byte of the UbxNavTime Msg.

We have configured the GPS to get a block each second. (Sol, Pollsh,
Dop, Velned and TimeGps)

I don't think the order of the single messages is specified and may
differ.

This delay is no indicator of the quality of the GPS measurement but the
performance of the OBS firmware.